Rename ExpireSnapshotsOptionsParser to ExecuteOptionsParser
The parser file is not specific to expire_snapshots; renaming it to
ExecuteOptionsParser so it can host option parsers for other Iceberg
EXECUTE commands in the future.

Made-with: Cursor
